Fusion Slam Done

+++ src/main/java/bgu/spl/mics/application/objects/CloudPoint.java
@@ -6,21 +6,21 @@ package bgu.spl.mics.application.objects;
  */
 public class CloudPoint {
 
-    private int x;  
-    private int y;  
+    private double x;  
+    private double y;  
 
-    public CloudPoint(int x, int y) {
+    public CloudPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     // Getters and Setters for x, y coordinates
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
